refactor(color-select-panel): linear-gradient renderless
